+++ src/Clue/GraphComposer/Command/Show.php
@@ -2,28 +2,22 @@
 
 namespace Clue\GraphComposer\Command;
 
-use Symfony\Component\Console\Input\InputOption;
-use Symfony\Component\Console\Input\InputArgument;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Clue\GraphComposer\Graph\GraphComposer;
 
-class Show extends Command
+class Show extends AbstractCommand
 {
     protected function configure()
     {
-        $this->setName('show')
-             ->setDescription('Show dependency graph image for given project directory')
-             ->addArgument('dir', InputArgument::OPTIONAL, 'Path to project directory to scan', '.')
-             ->addOption('format', null, InputOption::VALUE_REQUIRED, 'Image format (svg, png, jpeg)', 'svg')
-           /*->addOption('dev', null, InputOption::VALUE_NONE, 'If set, Whether require-dev dependencies should be shown') */;
+        parent::configure();
+        $this
+            ->setName('show')
+            ->setDescription('Show dependency graph image for given project directory');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $graph = new GraphComposer($input->getArgument('dir'));
-        $graph->setFormat($input->getOption('format'));
+        $graph = $this->createGraph($input);
         $graph->displayGraph();
 
         return 0;