+++ src/Clue/GraphComposer/Exclusion/Dependency/NoDevDependencyRule.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Clue\GraphComposer\Exclusion\Dependency;
+
+use JMS\Composer\Graph\DependencyEdge;
+
+class NoDevDependencyRule implements DependencyRule
+{
+    public function isExcluded(DependencyEdge $dependency)
+    {
+        if ($dependency->isDevDependency()) {
+            return true;
+        }
+
+        return false;
+    }
+}