+++ src/Clue/GraphComposer/Exclusion/Package/NoPhpExtensionRule.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Clue\GraphComposer\Exclusion\Package;
+
+use JMS\Composer\Graph\PackageNode;
+
+class NoPhpExtensionRule implements PackageRule
+{
+    public function isExcluded(PackageNode $package)
+    {
+        if ($package->isPhpExtension()) {
+            return true;
+        }
+
+        return false;
+    }
+}