+++ src/Clue/GraphComposer/Graph/GraphComposer.php
@@ -2,10 +2,15 @@
 
 namespace Clue\GraphComposer\Graph;
 
-use Fhaculty\Graph\Graph;
+use Clue\GraphComposer\Exclusion\Dependency\ChainedDependencyRule;
+use Clue\GraphComposer\Exclusion\Dependency\DependencyRule;
+use Clue\GraphComposer\Exclusion\Package\ChainedPackageRule;
+use Clue\GraphComposer\Exclusion\Package\PackageRule;
 use Fhaculty\Graph\Attribute\AttributeAware;
 use Fhaculty\Graph\Attribute\AttributeBagNamespaced;
+use Fhaculty\Graph\Graph;
 use Graphp\GraphViz\GraphViz;
+use JMS\Composer\Graph\PackageNode;
 
 class GraphComposer
 {
@@ -13,21 +18,24 @@ class GraphComposer
         'fillcolor' => '#eeeeee',
         'style' => 'filled, rounded',
         'shape' => 'box',
-        'fontcolor' => '#314B5F'
+        'fontcolor' => '#314B5F',
     );
 
     private $layoutVertexRoot = array(
-        'style' => 'filled, rounded, bold'
+        'style' => 'filled, rounded, bold',
     );
 
     private $layoutEdge = array(
         'fontcolor' => '#767676',
         'fontsize' => 10,
-        'color' => '#1A2833'
+        'color' => '#1A2833',
     );
 
     private $layoutEdgeDev = array(
-        'style' => 'dashed'
+        'style' => 'dashed',
+        'fontcolor' => '#767676',
+        'fontsize' => 10,
+        'color' => '#1A2833',
     );
 
     private $dependencyGraph;
@@ -38,68 +46,62 @@ class GraphComposer
     private $graphviz;
 
     /**
+     * The maximum depth of dependency to display.
      *
-     * @param string $dir
-     * @param GraphViz|null $graphviz
+     * @var int
      */
-    public function __construct($dir, GraphViz $graphviz = null)
-    {
+    private $maxDepth;
+
+    /**
+     * @var PackageRule
+     */
+    private $packageExclusionRule;
+
+    /**
+     * @var DependencyRule
+     */
+    private $dependencyExclusionRule;
+
+    public function __construct(
+        $dir,
+        GraphViz $graphviz = null,
+        PackageRule $packageExclusionRule = null,
+        DependencyRule $dependencyExclusionRule = null,
+        $maxDepth = PHP_INT_MAX
+    ) {
         if ($graphviz === null) {
             $graphviz = new GraphViz();
             $graphviz->setFormat('svg');
         }
+
+        if ($packageExclusionRule === null) {
+            $packageExclusionRule = new ChainedPackageRule();
+        }
+
+        if ($dependencyExclusionRule === null) {
+            $dependencyExclusionRule = new ChainedDependencyRule();
+        }
+
         $analyzer = new \JMS\Composer\DependencyAnalyzer();
         $this->dependencyGraph = $analyzer->analyze($dir);
         $this->graphviz = $graphviz;
+        $this->packageExclusionRule = $packageExclusionRule;
+        $this->dependencyExclusionRule = $dependencyExclusionRule;
+        $this->maxDepth = $maxDepth;
     }
 
     /**
-     *
-     * @param string $dir
      * @return \Fhaculty\Graph\Graph
      */
     public function createGraph()
     {
         $graph = new Graph();
 
-        foreach ($this->dependencyGraph->getPackages() as $package) {
-            $name = $package->getName();
-            $start = $graph->createVertex($name, true);
-
-            $label = $name;
-            if ($package->getVersion() !== null) {
-                $label .= ': ' . $package->getVersion();
-            }
-
-            $this->setLayout($start, array('label' => $label) + $this->layoutVertex);
-
-            foreach ($package->getOutEdges() as $requires) {
-                $targetName = $requires->getDestPackage()->getName();
-                $target = $graph->createVertex($targetName, true);
-
-                $label = $requires->getVersionConstraint();
-
-                $edge = $start->createEdgeTo($target);
-                $this->setLayout($edge, array('label' => $label) + $this->layoutEdge);
-
-                if ($requires->isDevDependency()) {
-                    $this->setLayout($edge, $this->layoutEdgeDev);
-                }
-            }
-        }
-
-        $root = $graph->getVertex($this->dependencyGraph->getRootPackage()->getName());
-        $this->setLayout($root, $this->layoutVertexRoot);
+        $drawnPackages = array();
+        $rootPackage = $this->dependencyGraph->getRootPackage();
+        $this->drawPackageNode($graph, $rootPackage, $drawnPackages, $this->layoutVertexRoot);
 
         return $graph;
-    }
-
-    private function setLayout(AttributeAware $entity, array $layout)
-    {
-        $bag = new AttributeBagNamespaced($entity->getAttributeBag(), 'graphviz.');
-        $bag->setAttributes($layout);
-
-        return $entity;
     }
 
     public function displayGraph()
@@ -114,6 +116,80 @@ class GraphComposer
         $graph = $this->createGraph();
 
         return $this->graphviz->createImageFile($graph);
+    }
+
+    private function drawPackageNode(
+        Graph $graph,
+        PackageNode $packageNode,
+        array &$drawnPackages,
+        array $layoutVertex = null,
+        $depth = 0
+    ) {
+        // the root package may not excluded
+        // beginning with $depth = 1 the packages are filtered using the exclude rule
+        if ($depth > 0 && $this->packageExclusionRule->isExcluded($packageNode)) {
+            return null;
+        }
+
+        $name = $packageNode->getName();
+        // ensure that packages are only drawn once
+        // if two packages in the tree require a package twice
+        // then this dependency does not need to be drawn twice
+        // and the vertex is returned directly (so an edge can be added)
+        if (isset($drawnPackages[$name])) {
+            return $drawnPackages[$name];
+        }
+
+        if ($depth > $this->maxDepth) {
+            return null;
+        }
+
+        if ($layoutVertex === null) {
+            $layoutVertex = $this->layoutVertex;
+        }
+
+        $vertex = $drawnPackages[$name] = $graph->createVertex($name, true);
+
+        $label = $name;
+        if ($packageNode->getVersion()) {
+            $label .= ': ' .$packageNode->getVersion();
+        }
+        $this->setLayout($vertex, array('label' => $label) + $layoutVertex);
+
+        // this foreach will loop over the dependencies of the current package
+        foreach ($packageNode->getOutEdges() as $dependency) {
+            if ($this->dependencyExclusionRule->isExcluded($dependency)) {
+                continue;
+            }
+
+            // never show dev dependencies of dependencies:
+            // they are not relevant for the current application and are ignored by composer
+            if ($depth > 0 && $dependency->isDevDependency()) {
+                continue;
+            }
+
+            $targetVertex = $this->drawPackageNode($graph, $dependency->getDestPackage(), $drawnPackages, null, $depth + 1);
+
+            // drawPackageNode will return null if the package should not be shown
+            // also the dependencies of a package will be only drawn if max depth is not reached
+            // this ensures that packages in a deeper level will not have any dependency
+            if ($targetVertex && $depth < $this->maxDepth) {
+                $label = $dependency->getVersionConstraint();
+                $edge = $vertex->createEdgeTo($targetVertex);
+                $layoutEdge = $dependency->isDevDependency() ? $this->layoutEdgeDev : $this->layoutEdge;
+                $this->setLayout($edge, array('label' => $label) + $layoutEdge);
+            }
+        }
+
+        return $vertex;
+    }
+
+    private function setLayout(AttributeAware $entity, array $layout)
+    {
+        $bag = new AttributeBagNamespaced($entity->getAttributeBag(), 'graphviz.');
+        $bag->setAttributes($layout);
+
+        return $entity;
     }
 
     public function setFormat($format)